+++ sentry/src/main/java/io/sentry/clientreport/IClientReportStorage.java
@@ -1,0 +1,11 @@
+package io.sentry.clientreport;
+
+import java.util.List;
+import org.jetbrains.annotations.ApiStatus;
+
+@ApiStatus.Internal
+public interface IClientReportStorage {
+  void addCount(ClientReportKey key, Long count);
+
+  List<DiscardedEvent> resetCountsAndGet();
+}