+++ sentry/src/test/java/io/sentry/SentryOptionsManipulator.kt
@@ -1,0 +1,12 @@
+package io.sentry
+
+import io.sentry.clientreport.IClientReportRecorder
+
+class SentryOptionsManipulator {
+
+    companion object {
+        fun setClientReportRecorder(options: SentryOptions, clientReportRecorder: IClientReportRecorder) {
+            options.clientReportRecorder = clientReportRecorder
+        }
+    }
+}